Renamed MutationInterface.
The new name is SingleComponentMutationInterface.

The name highlights the crucial fact that all methods implemented
on the class operate on *exactly one* component at a time.

This stands in contrast to other interfaces that operate
on *blocks* of components.

Future naming will clarify the distinction across all the selection
classes.