Corrected an obscure logical voice identification corner case.
QUESTION: which of the two expressions below is syntactic according to LilyPond?

LILYPOND EXPRESSION 1:

    {
        \context Staff = "foo" {
            \context Voice = "bar" {
                c'8 (
                cs'8
                d'8
                ef'8
            }
        }
        \context Staff = "foo" {
            \context Voice = "bar" {
                e'8
                f'8
                fs'8
                g'8 )
            }
        }
    }


LILYPOND EXPRESSION 2:

    {
        \context Staff = "foo" {
            c'8 (
            cs'8
            d'8
            ef'8
        }
        \context Staff = "foo" {
            e'8
            f'8
            fs'8
            g'8 )
        }
    }


ANSWER: only expression #1 is syntactic. Expression #2 generates an
unterminated slur error.

Why is this?

Expression #1 features two like-named staves with like-named
voices; this is enough for LilyPond to know that the notes in the entire
expression live in the same context and can therefore be spanned by the slur.

Expression #2 features two like-named staves with no explicit voices
given at all. In this situation LilyPond creates two voices implicitly
*and the identity of the two voices differ from each other*. This means
that the notes in expression #2 live in fundamentally different contexts
and can not be spanned.

The commit checked in here makes a three-line adjustment to the logical
voice determination logic implemented on the Parentage class.

Why did a number of test files change in this commit?

Because Abjad previously forbade the application of spanners to expressions
like #1 and also to expressions like #2.

Abjad now hews even more tightly to LilyPond and allows expressions like #1
while continuing to forbid expressions like #2.

The changed tests included in this commit reflect Abjad's new tolerance
for expressions like #1.