Assigned counttime status to different components.
MODEL. The following components are counttime components:

    * all leaves (Note, Rest, Chord, Skip)
    * Tuplet (and FixedDurationTuplet)
    * Measure
    * Cluster

All other components are not counttime components.

This commit implements an _is_counttime_component class attribute on Component,
set to false. Leaf, Tuplet, Measure and Cluster then set this class attribute
to true.

USAGE. The idea of counttime components fills a (lexical) gap in our
description of the system. Counttime components already form the basis
of many types of selection the code in the experimental branch. So it seems
right that the distinction provide by counttime status should be worked into
the mainline. This means that we'll eventually have to write docs explaining
the distinctin to end users. This also means that the criterion of
*spannability* can now be rendered much more explicitly: spanners should
probably only attach to counttime components (even though the current
implementation allows spanners to attach to contexts).

OBSERVATION. Note that saying that a component is a counttime component is
almost (but not quite) equivalent to saying that a component is a noncontext
component. Voices, staves, staff groups and scores are all contexts and none is
a counttime component. The confounding case is the unqualified Abjad Container:
the unqualified container is neither a counttime component nor a context. And
this actually makes sense: the fact that there's always been something vaguely
liminal about the unqualified container is probably explained by the fact that
containers have no notational reality in the score, which is probably what the
neither-counttime-nor-context status of containers points out.