Began UnincorporatedCounttimeComponentSelection integration.
leaftools.make_leaves() now returns a selection of unincorporated
leaves (instead of a built-in list).

All tests pass without modification and show that the selection
class functions as a drop-in replacement for the built-in lists
previously used as return types.

The other maker functions will integrate 
UnincorporatedCounttimeComponentSelection in future commits.