Unifying split functions. MODEL update included.
Here are the current split functions in the system:

    * componenttools.split_component_by_duration()
    * componenttoos.split_components_by_durations()
    * SingleComponentMutationInterface.split_in_halves()
    * SingleComponentMutationInterface.split_leaf_by_duration()
    * SingleComponentMutationInterface.split_leaf_by_durations()
    * containertools.split_container_at_index()
    * containertools.split_container_at_indices()

There are seven such functions.

The generalization of these functions is:

    * componenttoos.split_components_by_durations()

The other six functions are syntactic sugar around 
componenttoos.split_components_by_durations().

MODEL: this commit starts the process of split unification.
The following six functions are now deprecated:

    * componenttools.split_component_by_duration()
    * SingleComponentMutationInterface.split_in_halves()
    * SingleComponentMutationInterface.split_leaf_by_duration()
    * SingleComponentMutationInterface.split_leaf_by_durations()
    * containertools.split_container_at_index()
    * containertools.split_container_at_indices()

Use ...

    * componenttoos.split_components_by_durations()

... instead.

At the very end of the process it should be possible to rename
componenttoos.split_components_by_durations() to simply
componenttools.split() and to bind the method to
ContiguousComponentMutationInterface, or equivalent.

This commit reimplements SingleComponentMutationInterface.split_in_halves()
as a one-line wrapper around componenttools.split_components_by_durations();
all tests pass and this shows that the method is redundant.
The method will be removed in the following commit.