Adjusted return type of fancy copy function.
    * componenttools.copy_governed_component_subtree_from_offset_to()

The function now returns a tuplet instead of a leaf when a new
tuplet is created by slicing a leaf at a non-power-of-two offset relative
to the start of the leaf. (This is just a preliminary step prior
to deprecation.)

DEPRECATED: componenttools.copy_governed_component_subtree_from_offset_to()
is now deprecated. Why? Because the function implements a two-step process
equal to copy-then-trim against a component. The following commits will
rename the function, then reimplement it, and then remove it. Probably
very few users were using this function; those that were will use a two-step
copy-then-trim recipe instead.