Extended Menu._run() method.
The change causes menus to return the string 'user entered lone return'
any time the user types a lone return (ie, no input) while interacting
with a menu.

The change allows developers to implement a default behavior against
a menu.

This commit implements a default behavior against the SegmentPackageProxy
main menu: pressing nothing is the same as pressing 'e'; both drop
the user into an editing session with the segment definition module.

All that is needed to add a default to a menu is to extend the
_handle_main_menu_result() method of any interactive class like this:

    def _handle_main_menu_result(self, result):
        ...
        elif result == 'user entered lone return':
            self.do_default_action_here()
        ...