Extended score package proxy.
ScorePackageProxy now knows to look for an instrumentation.py module.

If an instrumentation.py module exists then the score package proxy
ignores any 'instrumentation' tag in the metadata module.

If an instrumentation.py module does not exist then the score package proxy
uses the 'instrumentation' tag in the metadata module.

Note that the use of an 'instrumentation' tag in the metadata module
is now deprecated; instrumentation info should be externalized to a
dedicated instrumentation.py module.