Extended ModuleProxy.
ModuleProxy now allows initialization by means of either a packagesystem
path or a filesystem path.

BACKGROUND: ModuleProxy previously insisted on initialization with a
packagesystem path only. This way of doing things will be deprecated
in the future. The future way of doing things will be to pass around
filesystem paths only and then convert filesystem paths to packagesystem
paths on an as-needed basis only.