Finished 2.13 reference manual.
BUILD MILESTONE:

As of this commit all 18 chapters of the reference manual have been edited for
correctness, edited for consistency and rebuilt. Some chapters have been
removed or moved to other sections of the docs. The remaining chapters are
these:

    1. Annotations
    2. Articulations
    3. Chords
    4. Containers
    5. I/O
    6. Instruments
    7. Lilypond command marks
    8. Lilypond comments
    9. Lilypond files
    10. Measures
    11. Named pitches
    12. Notes
    13. Rests
    14. Scores
    15. Spanners
    16. Staves
    17. Tuplets
    18. Voices

How complete are the chapters in the reference manual? Some chapters are more
complete than others. The standard for inclusion the 2.13 release is that all
of the information in each chapter must be correct and must interpret during
build. This means that chapters are allowed to be incomplete so long as the
information they do contain is correct.

Are these all of the reference manual chapters? No. We will definitely add in
more reference manual chapters in the 2.14 development cycle.

What's the basis of a chapter in the manual? Reference manul chapters are all
basically presentations of class API entries written in prose. The reference
manual entries are organized around objects; this differs from the tutorial
section of the docs (which are organized around tasks).

How will we know when we've written enough reference manual chapters in 2.14?
We haven't yet determined completion. But perhaps the answer is that we will
write one chapter for each class loaded into the global namespace when you
start Abjad. By that standard we are actually quite close:

    Chord
    Container
    Duration [entry needed]
    Fraction [entry needed]
    Measure
    Multiplier [entry needed]
    NamedPitch
    Note
    Offset [entry needed]
    Rest
    Score
    Staff
    Tuplet
    Voice

Perhaps all other documentation of this sort will be module-level
documentation.