Changed behavior of mathtools.cumulative_sums().
OLD:

    >>> mathtools.cumulative_sums([1, 2, 3, 4])
    [1, 3, 6, 10]

NEW:
    
    >>> mathtools.cumulative_sums([1, 2, 3, 4])
    [0, 1, 3, 6, 10]

mathtools.cumulative_sums() now behaves exactly like
mathtools.cumulative_sums_zero().

To retrieve the old behavior, use the new start keyword and set it to none:

    >>> mathtools.cumulative_sums([1, 2, 3, 4], start=None)
    [1, 3, 6, 10]

DEPRECATION: note that mathtools.cumulative_sums() and
mathtools.cumulative_sums_zero() will be collapsed in a coming commit.
The final function will be mathtools.cumulative_sums().
The behavior of the function will be the same as
mathtools.cumulative_sums_zero().

NOTE: if you're using mathtools.cumulative_sums() this commit just broke your
code.

(But note that the vast majority of usage is of
mathtools.cumulative_sums_zero(), which is fine as of this commit.)