Changed functionality of Spanner.attach() method.
Consider this:
   
   >>> staff = Staff("c'8 d'8 e'8 f'8")
   >>> beam = spannertools.BeamSpanner()
   >>> beam.attach(staff)

OLD: spanner is attached to the *leaves* in staff (but not to the staff)

NEW: spanner is attached directly to the staff.

If you want the spanner attached to the leaves, say this:

    >>> beam.attach(staff[:])

CHANGE: this is a significant change in spanner attachment behavior.
But highly unlikely to impact many users.

* * *

Removed mathtools.interpolate_*() functions.

Functionality no longer supported.

Doesn't need to be noted in the 2.14 change log.