Implementing mandatory explicit spanner attachment regime.
OLD:

    >>> staff = Staff("c'8 d'8 e'8 f'8")
    >>> beam = spannertools.BeamSpanner(staff)

NEW:

    >>> staff = Staff("c'8 d'8 e'8 f'8")
    >>> beam = spannertools.BeamSpanner()
    >>> beam.attach(staff)


EXPLANATION: we will soon enforce a two-step create-then-attach pattern
for working with spanners. This will be intermedial to the creation
of a top-level attach() method.

This commit brings all test collateral in the spannertools package up to date
with the new two-step pattern.

Future commits will make this pattern pervasive through the entire codebase.