Added top-level attach() function.
The function is currently housed in the scoretools package.

The function signature is this:

    attach(attachable, component_expression)

An attachable is either a mark or a spanner.

The function returns the attachable when the attachable is a mark.

The function returns none when the attachable is a spanner.

The return values of the function mirror the return values of
Mark.attach() and Spanner.attach().

The function comprises a single-line implementation:

    return attachable.attach(component_expression)

This commit does not yet integrate attach().

Future commits will integrate attach() and replace Mark.attach()
and Spanner.attach() with attach().