Extended FilesystemAssetWrangler to handle git's empty directory prohibition.
This commit adds two lines to the FilesystemAssetWrangler.

The two lines cause the FilesystemAssetWrangler to pass silently
when the storehouse directory for an asset can not be found.

Why is this necessary?

Consider the case of the Blue Example Score included as part of the
scoremanagertools package.

The score contains no stylesheets. This means that the blue_example_score
directory contained an empty blue_example_score/stylesheets directory when
managed by Subversion. But the directory no longer exists because Git doesn't
allow empty directories.

So there are two options:

    1. recreate the directory and add an empty .gitignore file
    2. teach parts of the score manager what to do when a directory goes
    missing

This commit does #2.

Perhaps we will eventually do #1, too. But #2 is enough to fix the failing
experimental doctests.

All experimental doctests pass as of this commit.