Fixed three selector fix-me items.
Background: selector methods responsible for selecting classes were failing.
The reason for this is that the score manager previously identified classes
as *directories*. This is no longer the case after package flattening.
Classes are now housed in files instead of directories.

This commit teaches the Selector to identify classes with files instead of
directories. Future commits will teach this same thing to more parts of the
system.