Rewired MultimeasureRest.
MultimeasureRest no longer inherits from Rest.

MultimeasureRest now inherits directory from Leaf.

It may seem that multimeasure rests are a type of rest.

But it's not clear that modeling with inheritance is warranted in this one
case.

MODEL: the leaf inheritance hierarhcy is flat as of this commit.

    Component
        Leaf
            Chord
            MultimeasureRest
            NaturalHarmonic
            Note
            Rest
            Skip

A future commit will add an ArtificialHarmonic class.

ArtificialHarmonic will also inherit directly from Leaf.