Removed Instrument from global namespace.
Instrument is a concrete class and can be used to create a DIY instrument.

In general we want to encourage the use of the prebuilt instruments in
instrumenttools. So instrumenttools.Violin() instead of
instrumenttools.Instrument('Violin', 'Vn.').

Instrument is removed as a way of directing users to the prebuilt classes in
instrumenttools.