Removing 'indicator' from codebase.
EXPLANATION: the term is used in a few places in the codebase. Usage isn't
rigorous. The term is more or less being used to mean 'thing' or 'object' or
'entity' or 'item': LogicalVoiceIndicator, ChordQulityIndicator,
SuspensionIndicator, OmissionIndicator, and so on. (All of the classes
featuring 'Indicator' in their name occur in tonalanalysistools.) While there's
nothing inherently problematic in this use of the term, we'll get more benefit
from putting the term into service in a rigorous way as a generalized
replacement for marks and context marks (and possibly spanners and anything
else that can be attached).

So this commit starts the process of removing 'indicator' from the codebase so
that the term can be reintroduced as a rigoorously defined replacement for
marks and context marks.