Constrained Component._get_marks().
The function now asserts that all marks found are ContextMark instances.

This is part of the refactoring designed ultimately to remove both Mark and
ContextMark from the system.

The constraint added in this commit makes sense because there are no more
classes in the system that inherit (directly) from Mark; all classes in the
system that inherit from Mark currently inherit from ContextMark; so it makes
sense that any items found by Component._get_marks() must be ContextMark
instances.