Made Tempo into an atom.
This one was a difficult commit.

One doctest continues to fail after a couple of hours of effort at debugging:

Quantizer.py line 145:

    Expected:
        \new Score <<
            \new Staff {
                \new Voice {
                    \tempo 4=60
                    c'4
                    cs'4
                    \tempo 4=120
                    d'2
                    ef'4 ~
                    \tempo 4=90
                    ef'8.
                    e'4 ~
                    e'16 ~
                    \times 2/3 {
                        \tempo 4=30
                        e'32
                        f'8.
                        fs'8 ~
                        fs'32 ~
                    }
                    \times 2/3 {
                        fs'32
                        g'8.
                        r8 ~
                        r32
                    }
                }
            }
        >>

    Got:
        \new Score <<
            \new Staff {
                    \new Voice {
                            \tempo 4=60
                            c'4
                            cs'4
                            \tempo 4=120
                            d'2
                            ef'4..
                            e'4 ~
                            e'16 ~
                            \times 2/3 {
                                    e'32
                                    f'8.
                                    fs'8 ~
                                    fs'32 ~
                            }
                            \times 2/3 {
                                    fs'32
                                    g'8.
                                    r8 ~
                                    r32
                            }
                    }
            }
        >>

The example is the example in the main API entry that uses a beatwise q-schema
to shape the rhythmic events. Eyeballing the output shows that two tempi are
missing and that (more mysteriously) the rhythmic spelling of the ef'4 and e'4
are different between the two sets of output; that is, there are actually a
different number of events between the two.

RECOMMENDATION: allow the test to continue to fail until ContextMark has been
completely removed from the system. This should only be another day or two
because the only classes that continue to inherit from ContextMark are
StaffChange and Instrument. A significant amount of moderately tricky code will
be completely removed when ContextMark exits and I think this will be the best
time to try to debug the example that fails here.