All documentationtools classes now initialize from empty input.
MILESTONE: all classes everywhere in the system now initialze either from empty
input or else with the _default_positional_argument_values class variable.

We may be able to eliminate _default_positional_argument_values, too.

In general these last commits have changed almost all class initializers in the
system to be keyword-only. The exception are the classes that already
implemented *args: the initializer signature of those classes have been left
as-is.