BuildDirectoryManager now strips \score block from segment .ly files.
It's more flexible to copy segment .ly files stripped of their \score blocks
into the build directory because it allows for both the horizontal
concatenation of like-named contexts as well as the original vertical
concatenation of scores.

To get the vertical concatenation layout in your *-score-segments.ly file:

    \score {
        \include path/to/segment/one
    }
    \score {
        \include path/to/segment/two
    }

To get the horizontal concatenation layout in your *-score-segments.ly file:

    \score {
        {
            \include path/to/segment/one
            \include path/to/segment/two
        }
    }

Note in the second example that the segment includes are wrapped into a block
by curly braces. \score blocks require a single input argument, and wrapping
multiple include statements together allows them to be treated as a single
object.