Removed "Designing 2.14" from in-conversation section of website.
The section remains in the PDF version of the 2.14 docs. But this is probably
OK: the section is effectively buried in the PDF but presented front and center
on the website.

This commit removes the content from the website. We can edit and turn into
prose and then feature on the front page of the website after 2.14 deploys.