Removed Spanner.pop() and Spanner.pop_left().
CHANGE: should be noted in the 2.15 change log.

CHANGE: all spanner modification methods are now deprecated. These are:

    * Spanner.append()
    * Spanner.append_left()
    * Spanner.extend()
    * Spanner.extend_left()
    * Spanner.fracture()
    * Spanner.fuse()

MODEL: these changes will transform spanners from implementing the interface of
a deque (double-ended queue) to that of an (immutable) tuple.

RELEASE MANAGEMENT: it may make sense for 2.15 to be a release that centers
primarily on spanners. GeneralizedBeam will replace Beam and the spanner
interface changes here will be complete. If we complete (at long last) the API
entries for all the classes in spannertools then it will make sense to build
2.15 to deploy just that collection of changes.