Updated docstrings.
Changed docstrings of boolean properties.

Old: r'''True when ...'''

New: r'''Is true when ...'''

DOC GUIDELINES: all docstrings should eventually begin with a third-person
singular verb in the simple present. Read-only properties should begin with
"Gets ...". Read / write properties should begin with "Gets and sets ...".
Functions and methods should begin with the 3S present form corresponding to
the verb that heads the name of the function or method. We're quite close to
this now.