Added GalleryInputBlock to systemtools.
The class models a block of gallery input.

Example:

    systemtools.GalleryInputBlock(
        input_={
            'denominator_multiplier_exponent': 0,
            'beam_each_cell': True,
            'beam_cells_together': False,
            },
        divisions=[
            (4, 8), (3, 4), (2, 4), (1, 16), (1, 16), (7, 8), (5, 16),
            ],
        )

The example here models a block of input suitable for the EvenRunRhythmMaker.
It's possible that other types of gallery input block will be implemented in
future. It's also possible that we may be able to get away with only a single
such class.