Improved rhythm-maker input storage of empty iterables.
All rhyhtm-makers now store empty iterable input as none rather than as an
empty tuple.

This means that foo=() and bar=() no longer show up in the repr and storage
format of rhythm-makers. This cleans up the repr and storage format of
rhythm-makers and makes it easier to see what has been configured on a
rhythm-maker after initialization.