Restructured GalleryInputBlock.
GalleryInputBlock now accepts a division_lists input parameter.
Pass in as many division lists as you want. The gallery input block will create
exactly one maker and then call the maker on all division lists.

This facilitates a manner of presentation that allows three or four different
two-system scores per rhythm-maker configuration.

The realization here is that the unit of display for the gallery will be the
page: one page will show exactly one rhythm-maker configuration.

Several dozen rhythm-maker configurations are likely in the gallery. This means
that several dozen pages are likely in the gallery.