Added should_scale_contents property to Measure.
The property is currently a stub.

The property will implement a boolean. When set to true the measure will behave
in the Ferneyhough way and implicitly prolate its contents. When set to false
the measure will not prolate its contents.

The system default at the moment is that all measures prolate their contents.

The new flag will make it possible for measures with nonbinary time signatures
to avoid prolating their contents for the first time.