Changed interpreter representation of contexts.
OLD:

    >>> Voice("c'8 d'8 e'8")
    Voice{3}

NEW:

    >>> Voice("c'8 d'8 e'8")
    Voice("c'8 d'8 e'8")

But note that this is still the case:

    >>> voice_1 = Voice("c'8 d'8 e'8")
    >>> voice_2 = Voice("f'8 g'8")
    >>> Staff([voice_1, voice_2])
    Staff<<2>>

The principle is that contexts now examine their contents before interpreter
representation creation: if all contents are leaves then contexts returns an
evaluable representation; otherwise contexts return the old-style abbreviated
(and nonevaluable) representation.

Doesn't need to be noted in the change log.