Changed the behavior of non-power-of-two measures.
Recall that measures now implement a read / write should_scale_contents
property. That property previously defaulted to true. As of this commit that
property now defaults to false.

What does this mean?

This means that all measures created in the system previously defaulted to
Ferneyhough-style behavior; the system default has changed; all measures
created in the system now default to a 'literal' reading of time signatures.
Users who want Ferneyhough-style behavior of measures should set
should_scale_contents to true. The property can be set either at or after
measure initialization.

CHANGE: this is an important change in system default behavior that should be
included in the 2.15 release notes.

NOTE: only music using non-power-of-two measures will be affected.