Renamed GalleryInputBlock.
The new name is GalleryInputSpecifier.

The class is now housed in rhythmmakertools instead of systemtools.

We can move the class back to systemtools when or if it makes sense. Right now
it makes sense to keep the class in rhythmmakertools because it is specific to
rhythm-maker input. And it is about to become even more specific to
rhythm-maker input over the next couple of commits.