Renamed `object_` to `subject` systemwide.
This prevents all but one "<autosummary>" error when building the docs. The
final error is caused by the name `input_` in `GalleryInputSpecifier`.

These problems are all due to the trailing underscore. Sphinx (sometimes, but
not always) interprets the trailing underscore as part of ReST syntax, and
strips it from the name under consideration. With the underscore removed,
Sphinx is unable to locate the name of the property or method (since the
non-underscore version doesn't exist).