Renamed implicit_scaling=False keyword. Note included.
The old name was should_scale_contents=False.

NOTE: the title line of this commit reverses the role of the objects supplied
to the verb 'rename'. We've previously authored commit message like this:

"Renamed should_scale_contents=False. The new name is implicit_scaling=False."

Here we write this:

"Renamed implicit_scaling=False. The old name was should_scale_contents=False."

I think this newer way will prove to be better because it forces the newest
information into the first line. And GitHub (and most other collaborative
systems) privilege the information in the first line. Makes scanning the commit
logs easier, is more likely to be read when reading commit messages on the
train, and so on.